+++ tests/test_asend.py
@@ -1,15 +1,17 @@
+import json
+import typing
+
 import requests
 import pytest
-import json
-
 import pytest_httpx
+
 from tg_api import tg_methods, tg_types
 
 
 @pytest.mark.anyio
 async def test_httpx_mocking(
     httpx_mock: pytest_httpx.HTTPXMock,
-    get_message_response: tg_methods.SendMessageResponse,
+    get_message_response: dict[str, typing.Any],
 ):
     tg_types.Chat.update_forward_refs()
     tg_types.Message.update_forward_refs()
@@ -35,7 +37,7 @@ async def test_httpx_mocking(
 @pytest.mark.anyio
 async def test_photo_request_mocking(
     httpx_mock: pytest_httpx.HTTPXMock,
-    get_photo_response: tg_methods.SendPhotoResponse,
+    get_photo_response: dict[str, typing.Any],
 ):
     tg_types.Chat.update_forward_refs()
     tg_types.Message.update_forward_refs()
@@ -89,7 +91,7 @@ async def test_photo_request_mocking(
 @pytest.mark.anyio
 async def test_document_request_mocking(
     httpx_mock: pytest_httpx.HTTPXMock,
-    get_document_response: tg_methods.SendDocumentResponse,
+    get_document_response: dict[str, typing.Any],
 ):
     tg_types.Chat.update_forward_refs()
     tg_types.Message.update_forward_refs()
@@ -151,3 +153,29 @@ async def test_document_request_mocking(
             response = await tg_request.asend()
             assert get_document_response == json.loads(json_payload)
             assert get_document_response == response.dict()
+
+
+@pytest.mark.anyio
+async def test_delete_message_request_mocking(
+    httpx_mock: pytest_httpx.HTTPXMock,
+    delete_message_response: dict[str, typing.Any],
+):
+    tg_types.Chat.update_forward_refs()
+    tg_types.Message.update_forward_refs()
+    httpx_mock.add_response(
+        url='https://api.telegram.org/bottoken/deleteMessage',
+        method='POST',
+        headers={
+            'content-type': 'application/json',
+            'accept': 'application/json',
+        },
+        json=delete_message_response,
+    )
+
+    async with tg_methods.AsyncTgClient.setup('token'):
+        tg_request = tg_methods.DeleteMessageRequest(chat_id=1234567890, message_id=12345)
+        json_payload = await tg_request.apost_as_json('deleteMessage')
+        response = await tg_request.asend()
+        assert isinstance(response, tg_methods.DeleteMessageResponse)
+        assert delete_message_response == json.loads(json_payload)
+        assert delete_message_response == response.dict()